refactor: OrderByValue에서 UI와 Domain로직 분리 및 함수명 변경
